Simplification of predicate elimination conditions
- predicate elimination is based on domain's extent rather than on data
  stored in parallel dimension map,
- add test case,

+++ csrc/index_compute.cpp
@@ -1403,7 +1403,7 @@ std::vector<Val*> Index::getGlobalProducerStridedIndices(
       strides[i] = IrBuilder::getItemExpr(
           IrBuilder::getAttrExpr(
               IrBuilder::metadataExpr(producer_tv), "alloc_stride"),
-          stride_i++);
+          (int64_t)stride_i++);
     }
   }
 
@@ -1759,7 +1759,7 @@ std::vector<Val*> Index::getStrides(TensorView* tv) {
       }
       strides[i] = IrBuilder::getItemExpr(
           IrBuilder::getAttrExpr(IrBuilder::metadataExpr(tv), "alloc_stride"),
-          stride_i++);
+          (int64_t)stride_i++);
     }
   }
